# Make EngineState clone cheaper with Arc on all of the heavy objects