--- a/crates/nu-protocol/src/ast/expression.rs
+++ b/crates/nu-protocol/src/ast/expression.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use serde::{Deserialize, Serialize};
 
 use super::{Argument, Expr, ExternalArgument, RecordItem};
@@ -325,7 +327,7 @@ impl Expression {
                 expr.replace_span(working_set, replaced, new_span);
             }
             Expr::Block(block_id) => {
-                let mut block = working_set.get_block(*block_id).clone();
+                let mut block = (**working_set.get_block(*block_id)).clone();
 
                 for pipeline in block.pipelines.iter_mut() {
                     for element in pipeline.elements.iter_mut() {
@@ -333,10 +335,10 @@ impl Expression {
                     }
                 }
 
-                *block_id = working_set.add_block(block);
+                *block_id = working_set.add_block(Arc::new(block));
             }
             Expr::Closure(block_id) => {
-                let mut block = working_set.get_block(*block_id).clone();
+                let mut block = (**working_set.get_block(*block_id)).clone();
 
                 for pipeline in block.pipelines.iter_mut() {
                     for element in pipeline.elements.iter_mut() {
@@ -344,7 +346,7 @@ impl Expression {
                     }
                 }
 
-                *block_id = working_set.add_block(block);
+                *block_id = working_set.add_block(Arc::new(block));
             }
             Expr::Binary(_) => {}
             Expr::Bool(_) => {}
@@ -429,7 +431,7 @@ impl Expression {
                 }
             }
             Expr::RowCondition(block_id) | Expr::Subexpression(block_id) => {
-                let mut block = working_set.get_block(*block_id).clone();
+                let mut block = (**working_set.get_block(*block_id)).clone();
 
                 for pipeline in block.pipelines.iter_mut() {
                     for element in pipeline.elements.iter_mut() {
@@ -437,7 +439,7 @@ impl Expression {
                     }
                 }
 
-                *block_id = working_set.add_block(block);
+                *block_id = working_set.add_block(Arc::new(block));
             }
             Expr::Table(headers, cells) => {
                 for header in headers {