--- a/crates/nu-protocol/src/engine/state_delta.rs
+++ b/crates/nu-protocol/src/engine/state_delta.rs
@@ -12,13 +12,13 @@ use crate::RegisteredPlugin;
 /// can be applied to the global state to update it to contain both previous state and the state held
 /// within the delta.
 pub struct StateDelta {
-    pub(super) files: Vec<(String, usize, usize)>,
-    pub(crate) file_contents: Vec<(Vec<u8>, usize, usize)>,
+    pub(super) files: Vec<(Arc<String>, usize, usize)>,
+    pub(crate) file_contents: Vec<(Arc<Vec<u8>>, usize, usize)>,
     pub(super) virtual_paths: Vec<(String, VirtualPath)>,
     pub(super) vars: Vec<Variable>,          // indexed by VarId
     pub(super) decls: Vec<Box<dyn Command>>, // indexed by DeclId
-    pub blocks: Vec<Block>,                  // indexed by BlockId
-    pub(super) modules: Vec<Module>,         // indexed by ModuleId
+    pub blocks: Vec<Arc<Block>>,             // indexed by BlockId
+    pub(super) modules: Vec<Arc<Module>>,    // indexed by ModuleId
     pub(super) usage: Usage,
     pub scope: Vec<ScopeFrame>,
     #[cfg(feature = "plugin")]
@@ -131,7 +131,7 @@ impl StateDelta {
         self.scope.pop();
     }
 
-    pub fn get_file_contents(&self) -> &[(Vec<u8>, usize, usize)] {
+    pub fn get_file_contents(&self) -> &[(Arc<Vec<u8>>, usize, usize)] {
         &self.file_contents
     }
 }