… crates/nu-protocol/src/engine/state_working_set.rs