@kubouch 2024-03-18
Interesting. I'd be willing to try it. Did you notice any runtime difference? Since this is intended as a perf optimization, it would be good to have some benchmarks. Also to make sure we don't regress, but it doesn't seem like we're introducing any overhead to the hot path.