@devyn 2024-03-18
@kubouch 

Yeah, I totally forgot to paste them :facepalm: 

## Before change

```
Timer precision: 30 ns
benchmarks                       fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ load_standard_lib             17.87 ms      │ 22.92 ms      │ 18.08 ms      │ 18.18 ms      │ 100     │ 100
├─ decoding_benchmarks                         │               │               │               │         │
│  ├─ json_decode                              │               │               │               │         │
│  │  ├─ (100, 5)                816.4 µs      │ 939.1 µs      │ 837.9 µs      │ 845.3 µs      │ 100     │ 100
│  │  ╰─ (10000, 15)             276.8 ms      │ 337.5 ms      │ 279.7 ms      │ 280.6 ms      │ 100     │ 100
│  ╰─ msgpack_decode                           │               │               │               │         │
│     ├─ (100, 5)                340.1 µs      │ 379.4 µs      │ 348.8 µs      │ 349.9 µs      │ 100     │ 100
│     ╰─ (10000, 15)             105 ms        │ 116.8 ms      │ 106.9 ms      │ 107.1 ms      │ 100     │ 100
├─ encoding_benchmarks                         │               │               │               │         │
│  ├─ json_encode                              │               │               │               │         │
│  │  ├─ (100, 5)                125.6 µs      │ 262.7 µs      │ 132.2 µs      │ 136.4 µs      │ 100     │ 100
│  │  ╰─ (10000, 15)             31.7 ms       │ 33.95 ms      │ 32.92 ms      │ 32.9 ms       │ 100     │ 100
│  ╰─ msgpack_encode                           │               │               │               │         │
│     ├─ (100, 5)                58.04 µs      │ 114.9 µs      │ 60.11 µs      │ 70.18 µs      │ 100     │ 100
│     ╰─ (10000, 15)             14.3 ms       │ 18.17 ms      │ 14.67 ms      │ 14.72 ms      │ 100     │ 100
├─ eval_benchmarks                             │               │               │               │         │
│  ├─ eval_default_config        4.343 ms      │ 15.27 ms      │ 4.51 ms       │ 4.85 ms       │ 100     │ 100
│  ╰─ eval_default_env           689.4 µs      │ 1.006 ms      │ 697.6 µs      │ 710.5 µs      │ 100     │ 100
├─ eval_commands                               │               │               │               │         │
│  ├─ each                                     │               │               │               │         │
│  │  ├─ 1                       3.106 ms      │ 5.146 ms      │ 3.164 ms      │ 3.214 ms      │ 100     │ 100
│  │  ├─ 5                       3.328 ms      │ 3.602 ms      │ 3.383 ms      │ 3.39 ms       │ 100     │ 100
│  │  ├─ 10                      3.59 ms       │ 4.222 ms      │ 3.676 ms      │ 3.708 ms      │ 100     │ 100
│  │  ├─ 100                     8.423 ms      │ 9.15 ms       │ 8.53 ms       │ 8.585 ms      │ 100     │ 100
│  │  ╰─ 1000                    56.87 ms      │ 64.93 ms      │ 62.56 ms      │ 62.22 ms      │ 100     │ 100
│  ├─ for_range                                │               │               │               │         │
│  │  ├─ 1                       3.053 ms      │ 5.401 ms      │ 3.181 ms      │ 3.275 ms      │ 100     │ 100
│  │  ├─ 5                       3.278 ms      │ 3.965 ms      │ 3.401 ms      │ 3.419 ms      │ 100     │ 100
│  │  ├─ 10                      3.533 ms      │ 4.02 ms       │ 3.655 ms      │ 3.658 ms      │ 100     │ 100
│  │  ├─ 100                     8.384 ms      │ 9.031 ms      │ 8.524 ms      │ 8.553 ms      │ 100     │ 100
│  │  ╰─ 1000                    56.99 ms      │ 66.22 ms      │ 62.42 ms      │ 61.99 ms      │ 100     │ 100
│  ├─ interleave                               │               │               │               │         │
│  │  ├─ 100                     1.989 ms      │ 3.294 ms      │ 2.103 ms      │ 2.168 ms      │ 100     │ 100
│  │  ├─ 1000                    11.26 ms      │ 15.39 ms      │ 12.66 ms      │ 13.2 ms       │ 100     │ 100
│  │  ╰─ 10000                   92.83 ms      │ 141.7 ms      │ 117.2 ms      │ 120.3 ms      │ 100     │ 100
│  ├─ interleave_with_ctrlc                    │               │               │               │         │
│  │  ├─ 100                     1.777 ms      │ 2.656 ms      │ 2.102 ms      │ 2.086 ms      │ 100     │ 100
│  │  ├─ 1000                    8.579 ms      │ 15.53 ms      │ 12.81 ms      │ 13.06 ms      │ 100     │ 100
│  │  ╰─ 10000                   108 ms        │ 145.6 ms      │ 130.3 ms      │ 129.3 ms      │ 100     │ 100
│  ├─ par_each_1t                              │               │               │               │         │
│  │  ├─ 1                       1.02 ms       │ 1.664 ms      │ 1.085 ms      │ 1.102 ms      │ 100     │ 100
│  │  ├─ 5                       1.254 ms      │ 1.416 ms      │ 1.318 ms      │ 1.319 ms      │ 100     │ 100
│  │  ├─ 10                      1.503 ms      │ 1.84 ms       │ 1.587 ms      │ 1.594 ms      │ 100     │ 100
│  │  ├─ 100                     6.758 ms      │ 8.777 ms      │ 7.431 ms      │ 7.42 ms       │ 100     │ 100
│  │  ╰─ 1000                    55.41 ms      │ 62.91 ms      │ 60.05 ms      │ 59.39 ms      │ 100     │ 100
│  ╰─ par_each_2t                              │               │               │               │         │
│     ├─ 1                       1.034 ms      │ 1.854 ms      │ 1.084 ms      │ 1.147 ms      │ 100     │ 100
│     ├─ 5                       1.162 ms      │ 1.404 ms      │ 1.209 ms      │ 1.221 ms      │ 100     │ 100
│     ├─ 10                      1.254 ms      │ 1.665 ms      │ 1.323 ms      │ 1.361 ms      │ 100     │ 100
│     ├─ 100                     4.005 ms      │ 5.285 ms      │ 4.573 ms      │ 4.535 ms      │ 100     │ 100
│     ╰─ 1000                    28.74 ms      │ 31.87 ms      │ 30.79 ms      │ 30.58 ms      │ 100     │ 100
╰─ parser_benchmarks                           │               │               │               │         │
   ├─ parse_default_config_file  3.208 ms      │ 6.518 ms      │ 3.476 ms      │ 3.662 ms      │ 100     │ 100
   ╰─ parse_default_env_file     558.5 µs      │ 791.4 µs      │ 567.5 µs      │ 582.7 µs      │ 100     │ 100
```

## After change

```
benchmarks                       fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ load_standard_lib             17.93 ms      │ 22.86 ms      │ 18.12 ms      │ 18.29 ms      │ 100     │ 100
├─ decoding_benchmarks                         │               │               │               │         │
│  ├─ json_decode                              │               │               │               │         │
│  │  ├─ (100, 5)                851.1 µs      │ 967 µs        │ 859.8 µs      │ 863.6 µs      │ 100     │ 100
│  │  ╰─ (10000, 15)             282.9 ms      │ 352.9 ms      │ 289.2 ms      │ 290.2 ms      │ 100     │ 100
│  ╰─ msgpack_decode                           │               │               │               │         │
│     ├─ (100, 5)                321.1 µs      │ 365.5 µs      │ 322 µs        │ 323.7 µs      │ 100     │ 100
│     ╰─ (10000, 15)             101.1 ms      │ 106.1 ms      │ 104.1 ms      │ 104 ms        │ 100     │ 100
├─ encoding_benchmarks                         │               │               │               │         │
│  ├─ json_encode                              │               │               │               │         │
│  │  ├─ (100, 5)                128.9 µs      │ 144.7 µs      │ 129.3 µs      │ 130.5 µs      │ 100     │ 100
│  │  ╰─ (10000, 15)             29.9 ms       │ 34.82 ms      │ 30.93 ms      │ 31.3 ms       │ 100     │ 100
│  ╰─ msgpack_encode                           │               │               │               │         │
│     ├─ (100, 5)                55.82 µs      │ 67.32 µs      │ 56.43 µs      │ 57.42 µs      │ 100     │ 100
│     ╰─ (10000, 15)             13.57 ms      │ 14.66 ms      │ 13.7 ms       │ 13.74 ms      │ 100     │ 100
├─ eval_benchmarks                             │               │               │               │         │
│  ├─ eval_default_config        4.209 ms      │ 15.78 ms      │ 4.229 ms      │ 4.456 ms      │ 100     │ 100
│  ╰─ eval_default_env           688.8 µs      │ 785.8 µs      │ 696.8 µs      │ 699.3 µs      │ 100     │ 100
├─ eval_commands                               │               │               │               │         │
│  ├─ each                                     │               │               │               │         │
│  │  ├─ 1                       166.3 µs      │ 272.1 µs      │ 167.9 µs      │ 170.3 µs      │ 100     │ 100
│  │  ├─ 5                       345.4 µs      │ 418.4 µs      │ 384.9 µs      │ 385 µs        │ 100     │ 100
│  │  ├─ 10                      635.9 µs      │ 719.9 µs      │ 655.9 µs      │ 658.9 µs      │ 100     │ 100
│  │  ├─ 100                     5.475 ms      │ 7.33 ms       │ 5.513 ms      │ 5.544 ms      │ 100     │ 100
│  │  ╰─ 1000                    53.89 ms      │ 61.87 ms      │ 58.48 ms      │ 57.15 ms      │ 100     │ 100
│  ├─ for_range                                │               │               │               │         │
│  │  ├─ 1                       147.1 µs      │ 307.4 µs      │ 176.8 µs      │ 216 µs        │ 100     │ 100
│  │  ├─ 5                       342.5 µs      │ 539 µs        │ 448.1 µs      │ 444.3 µs      │ 100     │ 100
│  │  ├─ 10                      617.8 µs      │ 831.1 µs      │ 657.2 µs      │ 684.9 µs      │ 100     │ 100
│  │  ├─ 100                     5.494 ms      │ 6.477 ms      │ 5.966 ms      │ 5.92 ms       │ 100     │ 100
│  │  ╰─ 1000                    53.84 ms      │ 59.47 ms      │ 57.75 ms      │ 57.59 ms      │ 100     │ 100
│  ├─ interleave                               │               │               │               │         │
│  │  ├─ 100                     1.224 ms      │ 1.637 ms      │ 1.339 ms      │ 1.38 ms       │ 100     │ 100
│  │  ├─ 1000                    10.54 ms      │ 14.29 ms      │ 12.28 ms      │ 12.37 ms      │ 100     │ 100
│  │  ╰─ 10000                   90.11 ms      │ 135.1 ms      │ 120.9 ms      │ 121.2 ms      │ 100     │ 100
│  ├─ interleave_with_ctrlc                    │               │               │               │         │
│  │  ├─ 100                     1.151 ms      │ 1.622 ms      │ 1.273 ms      │ 1.301 ms      │ 100     │ 100
│  │  ├─ 1000                    10.43 ms      │ 12.44 ms      │ 11.59 ms      │ 11.54 ms      │ 100     │ 100
│  │  ╰─ 10000                   101 ms        │ 138.2 ms      │ 121.2 ms      │ 122.4 ms      │ 100     │ 100
│  ├─ par_each_1t                              │               │               │               │         │
│  │  ├─ 1                       219.6 µs      │ 405.8 µs      │ 246.7 µs      │ 250.2 µs      │ 100     │ 100
│  │  ├─ 5                       446.5 µs      │ 939.9 µs      │ 512.7 µs      │ 539.8 µs      │ 100     │ 100
│  │  ├─ 10                      831.1 µs      │ 1.136 ms      │ 892.7 µs      │ 897.7 µs      │ 100     │ 100
│  │  ├─ 100                     5.819 ms      │ 6.832 ms      │ 6.221 ms      │ 6.208 ms      │ 100     │ 100
│  │  ╰─ 1000                    54.59 ms      │ 61.37 ms      │ 58.94 ms      │ 58.85 ms      │ 100     │ 100
│  ╰─ par_each_2t                              │               │               │               │         │
│     ├─ 1                       346.5 µs      │ 471.2 µs      │ 370.6 µs      │ 375.9 µs      │ 100     │ 100
│     ├─ 5                       342.9 µs      │ 641.1 µs      │ 383.4 µs      │ 389.2 µs      │ 100     │ 100
│     ├─ 10                      457.5 µs      │ 771.5 µs      │ 519.9 µs      │ 537.8 µs      │ 100     │ 100
│     ├─ 100                     3.193 ms      │ 3.506 ms      │ 3.367 ms      │ 3.359 ms      │ 100     │ 100
│     ╰─ 1000                    28.26 ms      │ 30.14 ms      │ 29.73 ms      │ 29.62 ms      │ 100     │ 100
╰─ parser_benchmarks                           │               │               │               │         │
   ├─ parse_default_config_file  3.098 ms      │ 6.874 ms      │ 3.124 ms      │ 3.532 ms      │ 100     │ 100
   ╰─ parse_default_env_file     549 µs        │ 643.6 µs      │ 560.1 µs      │ 563.9 µs      │ 100     │ 100
```